[Bronze V] Title: 이상한 기호, Time: 108 ms, Memory: 14252 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -14,10 +14,10 @@ public class Main {
 
     void solve() throws IOException {
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int a = Integer.parseInt(st.nextToken());
-        int b = Integer.parseInt(st.nextToken());
+        long a = Long.parseLong(st.nextToken());
+        long b = Long.parseLong(st.nextToken());
 
-        int c = (a + b) * (a - b);
+        long c = (a + b) * (a - b);
 
         bw.write(c + "");
     }
